requirements.txt
spotipy==2.19.0
ytmusicapi==0.19.5
pytube==11.0.1
mysql-connector-python==8.0.27
Pillow==8.4.0